Refactor into reusables
