+++ app/src/main/AndroidManifest.xml
@@ -16,11 +16,6 @@
         android:theme="@style/Theme.MyApplication"
         tools:targetApi="31">
         <activity
-            android:name="com.poste.DashboardActivity"
-            android:exported="false"
-            android:label="@string/title_activity_dashboard"
-            android:theme="@style/Theme.MyApplication" />
-        <activity
             android:windowSoftInputMode="adjustResize"
             android:name="com.poste.IntroActivity"
             android:exported="true"
