+++ app/src/main/java/com/poste/DashboardScreen.kt
@@ -7,6 +7,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
+import com.poste.reusables.Folder
+import java.time.LocalDate
 
 @Composable
 fun DashboardScreen(navController: NavHostController) {
@@ -15,7 +17,6 @@ fun DashboardScreen(navController: NavHostController) {
         modifier = Modifier.fillMaxSize(),
         color = MaterialTheme.colorScheme.background
     ) {
-        Folder("Test")
+        Folder("Test", 13, LocalDate.now())
     }
-
 }
